+++ lib/services/unity_hub_service.dart
@@ -1,7 +1,10 @@
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:tuple/tuple.dart';
+
 import '../data/exceptions.dart';
+import '../utils/system_info.dart';
 
 class UnityHubService {
   late String _unityHubExe;
@@ -22,19 +25,14 @@ class UnityHubService {
     }
     final lines =
         result.stdout.toString().split('\n').map((e) => e.trim()).toList();
-    final regex = RegExp(r'^(.*) , installed at (.*)$');
-    final entries = lines.where((l) => regex.hasMatch(l)).map((l) {
-      final match = regex.allMatches(l).toList();
-      final version = match[0][1]!;
-      final unityExe = match[0][2]!;
-      return MapEntry(version, unityExe);
-    });
-    return Map.fromEntries(entries);
+    final entries = UnityHubOutputParser.parseInstalledEditors(lines);
+    return entries;
   }
 
-  Future<Process> installUnity(
+  Future<Tuple2<Process, List<String>>> installUnity(
     String version,
     String changeset,
+    Architecture arch,
     List<String>? modules,
   ) async {
     final args = [
@@ -46,6 +44,12 @@ class UnityHubService {
       '-c',
       changeset,
     ];
+
+    // https://forum.unity.com/threads/install-apple-silicon-arm64-using-hub-cli.1423695/#post-9120544
+    if (Platform.isMacOS && arch != Architecture.x64) {
+      args.addAll(['--architecture', 'arm64']);
+    }
+
     if (Platform.isLinux) {
       args.removeAt(0);
     }
@@ -54,7 +58,8 @@ class UnityHubService {
       args.addAll(modules);
     }
 
-    return Process.start(_unityHubExe, args);
+    final process = await Process.start(_unityHubExe, args);
+    return Tuple2<Process, List<String>>(process, [_unityHubExe, ...args]);
   }
 
   Uri getWindowsInstallerUri() {
@@ -65,5 +70,19 @@ class UnityHubService {
   Uri getMacInstallerUri() {
     return Uri.parse(
         'https://public-cdn.cloud.unity3d.com/hub/prod/UnityHubSetup.dmg');
+  }
+}
+
+class UnityHubOutputParser {
+  /// Parses the output of `editors -i`.
+  static Map<String, String> parseInstalledEditors(List<String> lines) {
+    final regex = RegExp(r'^([^ ]*) .*, installed at (.*)$');
+    final entries = lines.where((l) => regex.hasMatch(l)).map((l) {
+      final match = regex.allMatches(l).toList();
+      final version = match[0][1]!;
+      final unityExe = match[0][2]!;
+      return MapEntry(version, unityExe);
+    });
+    return Map.fromEntries(entries);
   }
 }